+++ megatron/__init__.py
@@ -2,5 +2,5 @@ from . import layers
 from . import visuals
 from . import io
 from . import nodes
+from . import layertools
 from .pipeline import Pipeline, load_pipeline
-from .utils import helpers