+++ megatron/layers/core.py
@@ -1,7 +1,10 @@
 import inspect
 from ..nodes import InputNode, TransformationNode
-from ..nodes.wrappers import FeatureSet
 from .. import utils
+
+
+class Input(InputNode):
+    pass
 
 
 class Layer:
@@ -24,26 +27,20 @@ class Layer:
         self.kwargs = kwargs
         self.n_outputs = n_outputs
 
-    def _call_on_nodes(self, nodes, out_name):
-        """Create new node(s) by applying transformation to given nodes.
+    def __call__(self, nodes):
+        """Creates a TransformationNode associated with this Transformation and the given InputNodes.
 
-        Number of output nodes is defined by self.n_outputs.
-        If there's multiple output nodes, they're returned as a FeatureSet.
+        When running eagerly, perform a fit and transform.
 
         Parameters
         ----------
-        nodes : megatron.Node(s)
-            nodes given as input to the layer's transformation.
-        out_name : str or list of str (default: None)
-            name(s) to give the newly created node(s).
+        inbound_nodes : list of megatron.InputNode / megatron.TransformationNode
+            the input nodes, whose data are to be passed to transform_fn when run.
         """
-        if out_name is None and self.n_outputs > 1:
-            out_name = [None for i in range(self.n_outputs)]
-        elif len(utils.generic.listify(out_name)) != self.n_outputs:
-            raise ValueError("Number of names does not match number of outputs")
+        nodes = utils.generic.listify(nodes)
 
         if self.n_outputs > 1:
-            out_nodes = [TransformationNode(self, nodes, out_name[i], i)
+            out_nodes = [TransformationNode(self, nodes, i)
                         for i in range(self.n_outputs)]
             for node in nodes:
                 node.outbound_nodes += out_nodes
@@ -51,9 +48,9 @@ class Layer:
                 for out_node in out_nodes:
                     out_node.fit()
                     out_node.transform()
-            out = FeatureSet(out_nodes)
+            out = out_nodes
         else:
-            out_node = TransformationNode(self, nodes, out_name)
+            out_node = TransformationNode(self, nodes)
             for node in nodes:
                 node.outbound_nodes.append(out_node)
             if all(node.output is not None for node in nodes):
@@ -61,51 +58,6 @@ class Layer:
                 out_node.transform()
             out = out_node
         return out
-
-    def _call_on_feature_set(self, feature_set):
-        """Create one new TransformationNode for each node in given feature set.
-
-        Parameters
-        ----------
-        feature_set : megatron.FeatureSet
-            feature set to map the transformation onto.
-        """
-        new_nodes = [self._call_on_nodes([node], node.name) for node in feature_set.nodes]
-        if isinstance(new_nodes[0], list):
-            feature_sets = [FeatureSet([group[i] for group in new_nodes])
-                            for i in range(len(new_nodes[0]))]
-        else:
-            feature_sets = FeatureSet(new_nodes)
-        return FeatureSet(new_nodes)
-
-    def __call__(self, inbound_nodes, name=None):
-        """Creates a TransformationNode associated with this Transformation and the given InputNodes.
-
-        When running eagerly, perform a fit and transform.
-
-        There are 3 options for how the inputs are laid out:
-            - All nodes: simply use nodes as input.
-            - One FeatureSet: map the transformation to all nodes in the FeatureSet.
-            - Mix of nodes and FeatureSet: unroll FeatureSet(s) into nodes, making a flat list.
-
-        Parameters
-        ----------
-        inbound_nodes : list of megatron.InputNode / megatron.TransformationNode or megatron.FeatureSet
-            the input nodes, whose data are to be passed to transform_fn when run.
-        name : str or list of str (default: None)
-            name(s) to give the newly created node(s).
-        """
-        inbound_nodes = utils.generic.listify(inbound_nodes)
-        if len(inbound_nodes) == 1 and isinstance(inbound_nodes[0], FeatureSet):
-            if name:
-                raise ValueError("When transforming FeatureSet, name cannot be provided")
-            return self._call_on_feature_set(inbound_nodes[0])
-        elif any(isinstance(node, FeatureSet) for node in inbound_nodes):
-            nodes = [node.nodes if isinstance(node, FeatureSet) else node for node in inbound_nodes]
-            nodes = list(utils.generic.flatten(nodes))
-            return self._call_on_nodes(nodes, name)
-        else:
-            return self._call_on_nodes(inbound_nodes, name)
 
     def partial_fit(self, *inputs):
         """Updates metadata based on given batch of data or full dataset.