+++ megatron/nodes/__init__.py
@@ -1,3 +1,2 @@
 from .core import InputNode, TransformationNode
-from .wrappers import Input, FeatureSet
 from .fromfile import *