+++ megatron/utils/__init__.py
@@ -1,5 +1,4 @@
-from . import generic
-from . import helpers
 from . import pipeline
 from . import hash
 from . import errors
+from .generic import *